feat/component userSelectorを作成
